requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2
django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
markdown==3.4.1
pillow==9.2.0
psycopg2==2.9.4
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
